% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection.R
\name{detect_skin_tones}
\alias{detect_skin_tones}
\title{Simple person detection heuristic}
\usage{
detect_skin_tones(tl_images, downsample = 200)
}
\arguments{
\item{tl_images}{A tl_images tibble.}

\item{downsample}{Maximum side length for analysis. Default 200.}
}
\value{
The input tibble with added column:
\itemize{
\item \code{skin_tone_prop}: Proportion of pixels matching skin tone ranges.
}
}
\description{
A simple heuristic-based approach to estimate if people are present
based on skin tone detection.
}
\seealso{
Other detection: 
\code{\link{detect_faces}()}
}
\concept{detection}
