% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{extract_audio_rms}
\alias{extract_audio_rms}
\title{Extract simple audio loudness per frame}
\usage{
extract_audio_rms(tl_images, video_path = NULL, fps = NULL)
}
\arguments{
\item{tl_images}{A tl_images tibble.}

\item{video_path}{Path to video file.}

\item{fps}{Frames per second (required if no timing columns).}
}
\value{
The input tibble with added \code{audio_rms} column.
}
\description{
A lightweight version of \code{\link[=extract_audio_features]{extract_audio_features()}} that only computes
RMS loudness. Useful when you only need basic volume information.
}
\seealso{
Other audio: 
\code{\link{extract_audio_features}()}
}
\concept{audio}
