% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{extract_colourfulness}
\alias{extract_colourfulness}
\title{Compute colourfulness (M3 metric)}
\usage{
extract_colourfulness(tl_images, downsample = 200)
}
\arguments{
\item{tl_images}{A tl_images tibble.}

\item{downsample}{Maximum side length for analysis. Default 200.}
}
\value{
The input tibble with added column:
\itemize{
\item \code{colourfulness}: M3 metric value.
}
}
\description{
Compute the Hasler & Süsstrunk M3 colourfulness metric.
}
\references{
Hasler, D. and Süsstrunk, S. E. (2003). Measuring colorfulness in natural images.
}
\seealso{
Other color: 
\code{\link{extract_brightness}()},
\code{\link{extract_color_mean}()},
\code{\link{extract_color_median}()},
\code{\link{extract_color_mode}()},
\code{\link{extract_color_moments}()},
\code{\link{extract_color_variance}()},
\code{\link{extract_dominant_color}()},
\code{\link{extract_hue_histogram}()},
\code{\link{extract_saturation}()}
}
\concept{color}
