% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{frame_extract_dominant_color}
\alias{frame_extract_dominant_color}
\title{Extract dominant color}
\usage{
frame_extract_dominant_color(tl_frames, downsample = 100)
}
\arguments{
\item{tl_frames}{A tl_frames tibble.}

\item{downsample}{Maximum side length for analysis. Default 100.}
}
\value{
The input tibble with added columns:
\itemize{
\item \code{dominant_color_hex}: Hex code of the dominant color.
\item \code{dominant_color_r/g/b}: RGB components (0-255).
\item \code{dominant_color_proportion}: What fraction of pixels are this color.
}
}
\description{
Find the most common color in an image using k-means clustering.
}
\seealso{
Other color: 
\code{\link{frame_extract_brightness}()},
\code{\link{frame_extract_color_mean}()},
\code{\link{frame_extract_color_median}()},
\code{\link{frame_extract_color_mode}()},
\code{\link{frame_extract_color_moments}()},
\code{\link{frame_extract_color_variance}()},
\code{\link{frame_extract_colourfulness}()},
\code{\link{frame_extract_hue_histogram}()},
\code{\link{frame_extract_saturation}()}
}
\concept{color}
