% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fluency.R
\name{frame_extract_rule_of_thirds}
\alias{frame_extract_rule_of_thirds}
\title{Compute rule of thirds adherence}
\usage{
frame_extract_rule_of_thirds(tl_frames, downsample = 200)
}
\arguments{
\item{tl_frames}{A tl_frames tibble.}

\item{downsample}{Maximum side length for analysis. Default 200.}
}
\value{
The input tibble with added column:
\itemize{
\item \code{rule_of_thirds}: Score indicating adherence to rule of thirds (0-1).
}
}
\description{
Measure how well the image follows the rule of thirds composition.
}
\seealso{
Other fluency: 
\code{\link{frame_extract_center_bias}()},
\code{\link{frame_extract_fluency_metrics}()},
\code{\link{frame_extract_visual_complexity}()}
}
\concept{fluency}
