% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_frames.R
\name{is_tl_frames}
\alias{is_tl_frames}
\title{Check if object is tl_frames}
\usage{
is_tl_frames(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} is a tl_frames tibble, \code{FALSE} otherwise.
}
\description{
Check if object is tl_frames
}
\seealso{
Other io: 
\code{\link{print.tl_frames}()}
}
\concept{io}
