% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_images.R
\name{is_tl_images}
\alias{is_tl_images}
\title{Check if object is tl_images}
\usage{
is_tl_images(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} is a tl_images tibble, \code{FALSE} otherwise.
}
\description{
Check if object is tl_images
}
\seealso{
Other io: 
\code{\link{load_images}()},
\code{\link{print.tl_images}()}
}
\concept{io}
