% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_frames.R
\name{print.tl_frames}
\alias{print.tl_frames}
\title{Print method for tl_frames}
\usage{
\method{print}{tl_frames}(x, ...)
}
\arguments{
\item{x}{A tl_frames tibble.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The input \code{x} invisibly.
}
\description{
Print method for tl_frames
}
\seealso{
Other io: 
\code{\link{is_tl_frames}()}
}
\concept{io}
