% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_images.R
\name{print.tl_images}
\alias{print.tl_images}
\title{Print method for tl_images}
\usage{
\method{print}{tl_images}(x, ...)
}
\arguments{
\item{x}{A tl_images tibble.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
The input \code{x} invisibly.
}
\description{
Print method for tl_images
}
\seealso{
Other io: 
\code{\link{is_tl_images}()},
\code{\link{load_images}()}
}
\concept{io}
