% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylens-package.R
\docType{package}
\name{tinylens-package}
\alias{tinylens}
\alias{tinylens-package}
\title{tinylens: Tidy Image Analysis for Digital Humanities}
\description{
A tidy, pipeable toolkit for image-first analysis targeting digital
humanities and film studies. Provides functions for loading image collections,
extracting color metrics, composition analysis, face/object detection,
and neural embeddings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nabeelsiddiqui/tinylens}
  \item Report bugs at \url{https://github.com/nabeelsiddiqui/tinylens/issues}
}

}
\keyword{internal}
