% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{video_extract_frames_evenly}
\alias{video_extract_frames_evenly}
\title{Sample frames from video}
\usage{
video_extract_frames_evenly(
  video_path,
  n = 10,
  output_dir = tempdir(),
  format = "jpg"
)
}
\arguments{
\item{video_path}{Path to video file.}

\item{n}{Number of frames to extract (evenly spaced).}

\item{output_dir}{Directory to save extracted frames.}

\item{format}{Output image format. Default \code{"jpg"}.}
}
\value{
A tl_frames tibble of sampled frames.
}
\description{
Extract frames at regular intervals to create a summary of the video.
}
\seealso{
Other video: 
\code{\link{detect_shot_changes}()},
\code{\link{video_download}()},
\code{\link{video_extract_frames_by_seconds}()},
\code{\link{video_extract_keyframes}()},
\code{\link{video_extract_shots}()},
\code{\link{video_get_info}()}
}
\concept{video}
