% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylens-package.R
\docType{package}
\name{tidylens-package}
\alias{tidylens}
\alias{tidylens-package}
\title{tidylens: Tidy Image Analysis for Digital Humanities}
\description{
A tidy, pipeable toolkit for image-first analysis targeting digital
humanities and film studies. Provides functions for loading image collections,
extracting color metrics, composition analysis, face/object detection,
and neural embeddings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nabeelsiddiqui/tidylens}
  \item Report bugs at \url{https://github.com/nabeelsiddiqui/tidylens/issues}
}

}
\keyword{internal}
